:tractor: Move conversation files around
