+++ pkg/conversation/context.go
@@ -2,7 +2,6 @@ package conversation
 
 import (
 	"encoding/json"
-	"github.com/google/uuid"
 	"gopkg.in/yaml.v3"
 	"os"
 	"strings"
@@ -53,51 +52,6 @@ func loadFromJSONFile(filename string) ([]*Message, error) {
 	}
 
 	return messages, nil
-}
-
-type ManagerImpl struct {
-	Tree           *ConversationTree
-	ConversationID uuid.UUID
-}
-
-var _ Manager = (*ManagerImpl)(nil)
-
-type ManagerOption func(*ManagerImpl)
-
-func WithMessages(messages ...*Message) ManagerOption {
-	return func(m *ManagerImpl) {
-		m.AppendMessages(messages...)
-	}
-}
-
-func WithManagerConversationID(conversationID uuid.UUID) ManagerOption {
-	return func(m *ManagerImpl) {
-		m.ConversationID = conversationID
-	}
-}
-
-func NewManager(options ...ManagerOption) *ManagerImpl {
-	ret := &ManagerImpl{
-		ConversationID: uuid.Nil,
-		Tree:           NewConversationTree(),
-	}
-	for _, option := range options {
-		option(ret)
-	}
-
-	if ret.ConversationID == uuid.Nil {
-		ret.ConversationID = uuid.New()
-	}
-
-	return ret
-}
-
-func (c *ManagerImpl) GetConversation() Conversation {
-	return c.Tree.GetLeftMostThread(c.Tree.RootID)
-}
-
-func (c *ManagerImpl) GetMessage(ID NodeID) (*Message, bool) {
-	return c.Tree.GetMessage(ID)
 }
 
 func (c *ManagerImpl) AppendMessages(messages ...*Message) {
