+++ pkg/conversation/tree.go
@@ -2,26 +2,14 @@ package conversation
 
 import (
 	"encoding/json"
-	fmt "fmt"
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
 	"os"
-	"strings"
-	"time"
 )
 
-type ContentType string
+// SaveToFile saves the conversation tree to a JSON file.
 
-const (
-	ContentTypeChatMessage ContentType = "chat-message"
-)
-
-// NodeContent is an interface for different types of node content.
-type NodeContent interface {
-	ContentType() ContentType
-	String() string
-	View() string
-}
+// LoadFromFile loads the conversation tree from a JSON file.
 
 type NodeID uuid.UUID
 
@@ -40,111 +28,6 @@ func (id *NodeID) UnmarshalJSON(data []byte) error {
 
 func NewNodeID() NodeID {
 	return NodeID(uuid.New())
-}
-
-// Message represents a single message in the conversation.
-type Message struct {
-	ParentID   NodeID    `json:"parentID"`
-	ID         NodeID    `json:"id"`
-	Time       time.Time `json:"time"`
-	LastUpdate time.Time `json:"lastUpdate"`
-
-	Content  NodeContent            `json:"content"`
-	Metadata map[string]interface{} `json:"metadata"` // Flexible metadata field
-
-	Children []*Message `json:"children,omitempty"`
-}
-
-type MessageOption func(*Message)
-
-func WithMetadata(metadata map[string]interface{}) MessageOption {
-	return func(message *Message) {
-		message.Metadata = metadata
-	}
-}
-
-func WithTime(time time.Time) MessageOption {
-	return func(message *Message) {
-		message.Time = time
-	}
-}
-
-func WithParentID(parentID NodeID) MessageOption {
-	return func(message *Message) {
-		message.ParentID = parentID
-	}
-}
-
-func WithID(id NodeID) MessageOption {
-	return func(message *Message) {
-		message.ID = id
-	}
-}
-
-func NewMessage(content NodeContent, options ...MessageOption) *Message {
-	ret := &Message{
-		Content:    content,
-		ID:         NodeID(uuid.New()),
-		Time:       time.Now(),
-		LastUpdate: time.Now(),
-	}
-
-	for _, option := range options {
-		option(ret)
-	}
-
-	return ret
-}
-
-func NewChatMessage(role Role, text string, options ...MessageOption) *Message {
-	return NewMessage(&ChatMessageContent{
-		Role: role,
-		Text: text,
-	}, options...)
-}
-
-type Conversation []*Message
-
-type Role string
-
-const (
-	RoleSystem    Role = "system"
-	RoleAssistant Role = "assistant"
-	RoleUser      Role = "user"
-)
-
-type ChatMessageContent struct {
-	Role Role   `json:"role"`
-	Text string `json:"text"`
-}
-
-func (c *ChatMessageContent) ContentType() ContentType {
-	return ContentTypeChatMessage
-}
-
-func (c *ChatMessageContent) String() string {
-	return c.Text
-}
-
-func (c *ChatMessageContent) View() string {
-	// If we are markdown, add a newline so that it becomes valid markdown to parse.
-	if strings.HasPrefix(c.Text, "```") {
-		c.Text = "\n" + c.Text
-	}
-	return fmt.Sprintf("[%s]: %s", c.Role, strings.TrimRight(c.Text, "\n"))
-}
-
-var _ NodeContent = (*ChatMessageContent)(nil)
-
-func (mn *Message) MarshalJSON() ([]byte, error) {
-	type Alias Message
-	return json.Marshal(&struct {
-		ContentType ContentType `json:"contentType"`
-		*Alias
-	}{
-		ContentType: mn.Content.ContentType(),
-		Alias:       (*Alias)(mn),
-	})
 }
 
 // Intermediate representation for unmarshaling.
@@ -181,7 +64,15 @@ func (mn *Message) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// ConversationTree holds the entire conversation.
+// ConversationTree represents a tree-like structure for storing and managing conversation messages.
+//
+// The tree consists of nodes (messages) connected by parent-child links. These relationships are done
+// through the parent ID field in each message. The root node is the starting point of the conversation,
+// and each node can have multiple children. The tree allows for traversing the conversation in various ways.
+//
+// Node relationships are stored in the Message datastructure as `Children []*Message`.
+//
+// Each node has a unique ID, and the tree keeps track of the root node ID and the last inserted node ID.
 type ConversationTree struct {
 	Nodes  map[NodeID]*Message
 	RootID NodeID
@@ -197,7 +88,9 @@ func NewConversationTree() *ConversationTree {
 
 var NullNode NodeID = NodeID(uuid.Nil)
 
-// InsertMessages adds a new message to the conversation tree.
+// InsertMessages adds new messages to the conversation tree.
+// It updates the root ID if the tree is empty and sets the last inserted node ID.
+// If a message has a parent ID that exists in the tree, it is added as a child of that parent node.
 func (ct *ConversationTree) InsertMessages(msgs ...*Message) {
 	for _, msg := range msgs {
 		ct.Nodes[msg.ID] = msg
@@ -212,6 +105,9 @@ func (ct *ConversationTree) InsertMessages(msgs ...*Message) {
 	}
 }
 
+// AttachThread attaches a conversation thread to a specified parent message.
+// It updates the parent IDs of the messages in the thread to link them to the parent message.
+// The last message in the thread becomes the new last inserted node ID.
 func (ct *ConversationTree) AttachThread(parentID NodeID, thread Conversation) {
 	for _, msg := range thread {
 		msg.ParentID = parentID
@@ -228,10 +124,16 @@ func (ct *ConversationTree) AttachThread(parentID NodeID, thread Conversation) {
 	}
 }
 
+// AppendMessages appends a conversation thread to the end of the tree.
+// It attaches the thread to the last inserted node in the tree, making it the parent of the thread.
+// The messages in the thread are inserted as nodes, extending the parent-child chain.
 func (ct *ConversationTree) AppendMessages(thread Conversation) {
 	ct.AttachThread(ct.LastID, thread)
 }
 
+// PrependThread prepends a conversation thread to the beginning of the tree.
+// It updates the root ID to the first message in the thread and adjusts the parent-child relationships accordingly.
+// The previous root node becomes a child of the new root node.
 func (ct *ConversationTree) PrependThread(thread Conversation) {
 	prevRootID := ct.RootID
 	newRootID := NullNode
@@ -253,7 +155,8 @@ func (ct *ConversationTree) PrependThread(thread Conversation) {
 	}
 }
 
-// FindSiblings returns the IDs of all sibling messages.
+// FindSiblings returns the IDs of all sibling messages for a given message ID.
+// Sibling messages are the nodes that share the same parent as the given message.
 func (ct *ConversationTree) FindSiblings(id NodeID) []NodeID {
 	node, exists := ct.Nodes[id]
 	if !exists {
@@ -275,7 +178,8 @@ func (ct *ConversationTree) FindSiblings(id NodeID) []NodeID {
 	return siblings
 }
 
-// FindChildren returns the IDs of all child messages.
+// FindChildren returns the IDs of all child messages for a given message ID.
+// Child messages are the nodes directly connected to the given message as its children.
 func (ct *ConversationTree) FindChildren(id NodeID) []NodeID {
 	node, exists := ct.Nodes[id]
 	if !exists {
@@ -290,7 +194,9 @@ func (ct *ConversationTree) FindChildren(id NodeID) []NodeID {
 	return children
 }
 
-// GetConversationThread returns the linear conversation thread starting from a given child ID.
+// GetConversationThread returns the linear conversation thread starting from a given message ID.
+// It traverses the tree upwards, following the parent links, to retrieve the complete thread.
+// The returned conversation is a linear sequence of messages from the root to the given message.
 func (ct *ConversationTree) GetConversationThread(id NodeID) Conversation {
 	var thread Conversation
 	for uuid.UUID(id) != uuid.Nil {
@@ -304,8 +210,9 @@ func (ct *ConversationTree) GetConversationThread(id NodeID) Conversation {
 	return thread
 }
 
-// GetLeftMostThread returns the thread starting from id by always chosing the first sibling
-// in the tree.
+// GetLeftMostThread returns the thread starting from a given message ID by always choosing the first child.
+// It traverses the tree downwards, selecting the leftmost child at each level, until a leaf node is reached.
+// The returned conversation is a linear sequence of messages from the given message to the leftmost leaf.
 func (ct *ConversationTree) GetLeftMostThread(id NodeID) Conversation {
 	var thread Conversation
 	for id != NullNode {
@@ -341,7 +248,8 @@ func (ct *ConversationTree) LoadFromFile(filename string) error {
 	return json.Unmarshal(data, ct)
 }
 
-func (ct *ConversationTree) GetMessage(id NodeID) (*Message, bool) {
+// GetMessageByID returns a message by its ID from the conversation tree.
+func (ct *ConversationTree) GetMessageByID(id NodeID) (*Message, bool) {
 	ret, exists := ct.Nodes[id]
 
 	return ret, exists
